🚚 move user folder into core/user
