+++ src/core/user/types/user-output.ts
